+++ src/model/date/date.h
@@ -8,7 +8,7 @@
 #include <iostream>
 #include <string>
 
-#include "../exception/date_exception.h"
+#include "exception/date_exception.h"
 
 class Date{
 public:
