+++ src/model/product/product.h
@@ -9,7 +9,7 @@
 #include <map>
 #include <set>
 
-#include "../exception/product_exception.h"
+#include "exception/product_exception.h"
 
 enum class CakeCategory {
     GENERAL, PIE, SPONGE, PUFF_PASTRY, CRUNCHY
